+++ spec/lib/yamlfile_spec.rb
@@ -10,12 +10,13 @@ describe RailsI18nterface::Yamlfile do
 
   describe 'write' do
     before(:each) do
-      @file_path = File.join(File.dirname(__FILE__), 'files', 'en.yml')
-      @file = RailsI18nterface::Yamlfile.new(@file_path)
+      @root_dir = File.expand_path(File.join('..', '..', '..', 'spec', 'internal'), __FILE__)
+      @file_path = File.join(@root_dir, 'config', 'locales', 'en.yml')
+      @file = RailsI18nterface::Yamlfile.new(@root_dir, :en)
     end
 
     after(:each) do
-      FileUtils.rm(@file_path)
+      FileUtils.rm(@file_path) if File.exists? @file_path
     end
 
     it 'writes all I18n messages for a locale to YAML file' do
